fix: replace old example
